linux: Override DIR_APP_DATA before loading .js
